[HOT-FIX[] corrected a lot of implicit integer conversion warnings
Former-commit-id: 6a354742bc66bfb138902bc1c686ee064db08b63
Former-commit-id: 2e599757566077f477dc811a0cc73879708fe4bf

+++ src/architecture/utilities/BSpline.cpp
@@ -96,13 +96,13 @@ OutputDataSet::~OutputDataSet()
 /*! This method returns x, xDot and xDDot at the desired input time T */
 void OutputDataSet::getData(double T, double x[3], double xDot[3], double xDDot[3])
 {
-    int N = this->T.size()-1;
+    int N = (int) this->T.size() - 1;
     double Ttot = this->T[N];
 
     // if T < Ttot calculalte the values
     if (T <= Ttot) {
         double t = T / Ttot;
-        int Q = this->C1.size();
+        int Q = (int) this->C1.size();
         Eigen::VectorXd NN(Q), NN1(Q), NN2(Q);
         basisFunction(t, this->U, Q, this->P, &NN[0], &NN1[0], &NN2[0]);
         x[0] = NN.dot(this->C1);
@@ -133,13 +133,13 @@ void OutputDataSet::getData(double T, double x[3], double xDot[3], double xDDot[
 /*! It is designed to be accessible from Python */
 double OutputDataSet::getStates(double T, int derivative, int index)
 {
-    int N = this->T.size()-1;
+    int N = (int) this->T.size()-1;
     double Ttot = this->T[N];
 
     // if T < Ttot calculalte the values
     if (T <= Ttot) {
         double t = T / Ttot;
-        int Q = this->C1.size();
+        int Q = (int) this->C1.size();
         Eigen::VectorXd NN(Q), NN1(Q), NN2(Q);
         basisFunction(t, this->U, Q, this->P, &NN[0], &NN1[0], &NN2[0]);
